cloud_functions/iss_api_get_loc_fact/requirements.txt
Flask>=2.2.5
Werkzeug>=3.0.6
functions-framework==3.*
google-cloud-secret-manager==2.16.4
google-cloud-storage==2.14.0
google-cloud-aiplatform>=1.38.0
